// frontend/src/app/rest-api/types.gen.ts
export type SigninRequest = {
    username: string;
    password: string;
};

export type SigninResponse = {
    message?: string;
    token?: string;
    username?: string;
    firstName?: string;
    lastName?: string;
    role?: string;
};

export type SignupRequest = {
    firstName: string;
    lastName: string;
    username: string;
    password: string;
};

export type SignupResponse = {
    message?: string;
    userId?: string;
};

export type ChangePasswordRequest = {
    username: string;
    oldPassword: string;
    newPassword: string;
};
// ...
export type PostAuthSigninData = {
    requestBody: SigninRequest;
};

export type PostAuthSigninResponse = SigninResponse;

export type PostAuthSignupData = {
    requestBody: SignupRequest;
};

export type PostAuthSignupResponse = SignupResponse;

export type PostAuthChangePasswordData = {
    requestBody: ChangePasswordRequest;
};